agents/chatbot/requirements.txt
streamlit>=1.31.0
langchain>=0.1.9
langchain-community>=0.0.24
langchain-huggingface>=0.0.2
langchain-chroma>=0.2.5
pypdf>=3.17.4
chromadb>=0.4.22
sentence-transformers>=2.2.2
groq>=0.4.0
langchain-groq>=0.1.3
python-dotenv>=1.0.1